📦 factor out generation of product codes into central src/mqt/qecc/codes directory
