+++ src/mqt/qecc/codes/constructions/lifted_product/lp_l=16_hx.npz
