+++ src/mqt/qecc/codes/constructions/lifted_product/lp_l=21_hz.npz
