+++ src/mqt/qecc/codes/constructions/lifted_product/lp_l=30_hz.npz
